[APPC-3541] separated the remaining possible ext utils
