+++ app/src/main/java/com/asfoundation/wallet/billing/address/BillingAddressRepository.kt
@@ -1,7 +1,7 @@
 package com.asfoundation.wallet.billing.address
 
-import com.asfoundation.wallet.util.component6
-import com.asfoundation.wallet.util.guardLet
+import com.appcoins.wallet.core.utils.common.extensions.component6
+import com.appcoins.wallet.core.utils.common.extensions.guardLet
 import com.appcoins.wallet.sharedpreferences.SecurePreferencesDataSource
 import javax.inject.Inject
