+++ app/src/main/java/com/asfoundation/wallet/eskills/withdraw/repository/WithdrawApiMapper.kt
@@ -3,7 +3,7 @@ package com.asfoundation.wallet.eskills.withdraw.repository
 import com.appcoins.wallet.billing.util.isNoNetworkException
 import com.asfoundation.wallet.eskills.withdraw.domain.FailedWithdraw
 import com.asfoundation.wallet.eskills.withdraw.domain.WithdrawResult
-import com.asfoundation.wallet.util.getMessage
+import com.appcoins.wallet.core.utils.common.extensions.getMessage
 import com.google.gson.Gson
 import com.google.gson.annotations.SerializedName
 import retrofit2.HttpException
