+++ app/src/main/java/com/asfoundation/wallet/main/nav_bar/NavBarFragment.kt
@@ -16,9 +16,9 @@ import by.kirich1409.viewbindingdelegate.viewBinding
 import com.asf.wallet.R
 import com.asf.wallet.databinding.NavBarFragmentBinding
 import com.appcoins.wallet.ui.arch.SingleStateFragment
+import com.appcoins.wallet.ui.common.createColoredString
+import com.appcoins.wallet.ui.common.setTextFromColored
 import com.asfoundation.wallet.main.MainActivity
-import com.asfoundation.wallet.util.createColoredString
-import com.asfoundation.wallet.util.setTextFromColored
 import com.asfoundation.wallet.viewmodel.BasePageViewFragment
 import com.google.android.material.bottomnavigation.BottomNavigationItemView
 import com.google.android.material.bottomnavigation.BottomNavigationMenuView
