+++ app/src/main/java/com/asfoundation/wallet/subscriptions/UserSubscriptionsMapper.kt
@@ -3,7 +3,7 @@ package com.asfoundation.wallet.subscriptions
 import com.appcoins.wallet.bdsbilling.SubscriptionSubStatus
 import com.appcoins.wallet.bdsbilling.subscriptions.UserSubscriptionsListResponse
 import com.asfoundation.wallet.util.Period
-import com.asfoundation.wallet.util.isNoNetworkException
+import com.appcoins.wallet.core.utils.common.extensions.isNoNetworkException
 import java.text.SimpleDateFormat
 import java.util.*
 import javax.inject.Inject
