+++ app/src/main/java/com/asfoundation/wallet/subscriptions/cancel/SubscriptionCancelPresenter.kt
@@ -1,7 +1,7 @@
 package com.asfoundation.wallet.subscriptions.cancel
 
 import com.asfoundation.wallet.subscriptions.UserSubscriptionsInteractor
-import com.asfoundation.wallet.util.isNoNetworkException
+import com.appcoins.wallet.core.utils.common.extensions.isNoNetworkException
 import io.reactivex.Observable
 import io.reactivex.Scheduler
 import io.reactivex.disposables.CompositeDisposable
