+++ app/src/main/java/com/asfoundation/wallet/ui/common/ViewExt.kt
@@ -2,7 +2,7 @@ package com.asfoundation.wallet.ui.common
 
 import android.view.View
 import android.view.ViewGroup
-import com.asfoundation.wallet.util.convertDpToPx
+import com.appcoins.wallet.ui.common.convertDpToPx
 
 fun View.setMargins(startMarginDp: Int? = null,
                     topMarginDp: Int? = null,
