+++ app/src/main/java/com/asfoundation/wallet/ui/iab/payments/carrier/confirm/CarrierFeeFragment.kt
@@ -17,7 +17,7 @@ import com.asfoundation.wallet.ui.iab.IabView
 import com.appcoins.wallet.core.utils.common.CurrencyFormatUtils
 import com.appcoins.wallet.core.utils.common.WalletCurrency
 import com.appcoins.wallet.core.utils.common.getStringSpanned
-import com.asfoundation.wallet.util.withNoLayoutTransition
+import com.appcoins.wallet.ui.common.withNoLayoutTransition
 import com.asfoundation.wallet.viewmodel.BasePageViewFragment
 import com.jakewharton.rxbinding2.view.RxView
 import dagger.hilt.android.AndroidEntryPoint
