+++ app/src/main/java/com/asfoundation/wallet/ui/iab/payments/carrier/verify/CarrierVerifyPresenter.kt
@@ -11,7 +11,7 @@ import com.appcoins.wallet.core.utils.common.CurrencyFormatUtils
 import com.appcoins.wallet.ui.common.StringProvider
 import com.appcoins.wallet.core.utils.common.WalletCurrency
 import com.appcoins.wallet.core.utils.common.applicationinfo.ApplicationInfoProvider
-import com.asfoundation.wallet.util.safeLet
+import com.appcoins.wallet.core.utils.common.extensions.safeLet
 import io.reactivex.Completable
 import io.reactivex.Observable
 import io.reactivex.Scheduler
