+++ app/src/main/java/com/asfoundation/wallet/ui/transact/TransferFragmentPresenter.kt
@@ -8,7 +8,7 @@ import com.asfoundation.wallet.ui.transact.TransferFragmentView.TransferData
 import com.appcoins.wallet.core.utils.common.CurrencyFormatUtils
 import com.asfoundation.wallet.util.QRUri
 import com.appcoins.wallet.core.utils.common.WalletCurrency
-import com.asfoundation.wallet.util.isNoNetworkException
+import com.appcoins.wallet.core.utils.common.extensions.isNoNetworkException
 import com.asfoundation.wallet.wallets.usecases.GetWalletInfoUseCase
 import com.google.android.gms.common.api.CommonStatusCodes
 import com.google.android.gms.vision.barcode.Barcode
