+++ core/utils/common/src/main/java/com/appcoins/wallet/core/utils/common/extensions/NetworkExt.kt
@@ -1,0 +1,17 @@
+package com.appcoins.wallet.core.utils.common.extensions
+
+import retrofit2.HttpException
+import java.io.IOException
+
+fun Throwable?.isNoNetworkException(): Boolean {
+  return this != null && (this is IOException || this.cause != null && this.cause is IOException)
+}
+
+fun HttpException.getMessage(): String {
+  val reader = this.response()
+    ?.errorBody()
+    ?.charStream()
+  val message = reader?.readText()
+  reader?.close()
+  return if (message.isNullOrBlank()) message() else message
+}
